requirements.txt
# PostgreSQL async driver
asyncpg==0.29.0
python-telegram-bot==20.7
yfinance==1.2.0
pandas==2.2.2
numpy==1.26.4
python-dotenv==1.0.1
openai==1.35.3
fastapi==0.110.0
uvicorn==0.29.0
